+++ include/core/hash/hasher.h
@@ -14,7 +14,7 @@ template<class T>
 requires std::is_pod_v<T>
 struct hasher<T> {
     using U = std::decay_t<T>;
-    size_t operator()(const T& value) const noexcept {
+    uint64 operator()(const T& value) const noexcept {
 	return mixer(std::hash<U>{}(value), 32);
     }
 };
@@ -22,8 +22,8 @@ struct hasher<T> {
 template<class T>
 requires core::mp::is_same_template_v<T, std::pair>
 struct hasher<T> {
-    size_t operator()(const T& value) const noexcept {
-	size_t v{0};
+    uint64 operator()(const T& value) const noexcept {
+	uint64 v{0};
 	combine(v, hasher<typename T::first_type>{}(value.first));
 	combine(v, hasher<typename T::second_type>{}(value.second));
 	return v;
@@ -37,8 +37,8 @@ requires (core::mp::is_same_template_v<T, std::vector> or
 	  core::mp::is_same_template_v<T, std::map> or
 	  core::mp::is_same_template_v<T, std::set>)
 struct hasher<T> {
-    size_t operator()(const T& container) const noexcept {
-	size_t v{0};
+    uint64 operator()(const T& container) const noexcept {
+	uint64 v{0};
 	for (const auto& element : container)
 	    combine(v, hasher<std::decay_t<decltype(element)>>{}(element));
 	return v;
